Trigger funcs (#405)
* remove old tools, fix warnings

* almost match si_handle_print_debug_var

* si_goto_end_loop and warnings

* is_trigger_bound and create_trigger

* Cleanup + matching 2 non-matchings

* update_triggers (trigger.c complete)

* fix warnings

* PR comments, m2ctx.py fix

+++ src/hud_element.h
@@ -232,6 +232,6 @@ void set_hud_element_transform_rotation_pivot(s32 id, s32 dx, s32 dy);
 
 void copy_world_hud_element_ref_to_battle(s32 worldID, s32 battleID);
 
-void set_hud_element_nonworld_cache(void* base, size_t size);
+void set_hud_element_nonworld_cache(void* base, s32 size);
 
 #endif
